pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "digitaltap-ai"
version = "0.1.0"
description = "Save 40%+ on cloud compute with local AI agents"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "Digital Tap AI", email = "oss@digitaltap.ai"},
]
keywords = ["cloud", "cost-optimization", "ai-agents", "databricks", "aws", "ollama"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
databricks = ["databricks-sdk>=0.20.0"]
aws = ["boto3>=1.34.0"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.23.0", "ruff>=0.1.0"]

[project.scripts]
digitaltap = "digitaltap.cli:main"

[project.urls]
Homepage = "https://digitaltap.ai"
Repository = "https://github.com/digitaltap/digitaltap-ai-oss"
Documentation = "https://github.com/digitaltap/digitaltap-ai-oss#readme"

[tool.setuptools.packages.find]
include = ["digitaltap*"]

[tool.ruff]
line-length = 100
target-version = "py310"
